backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
drf-spectacular>=0.28
